pass recovery check

+++ src/storage/memory_storage.h
@@ -17,7 +17,7 @@ public:
         size_t res_num = 0;
         res = malloc(128 * 2000);
 
-        for (int i = 0; i < datas.size(); ++i) {
+        for (size_t i = 0; i < datas.size(); ++i) {
             switch (where_column) {
             case 0:
                 is_equal = memcmp(column_key, &datas[i].id, column_key_len) == 0;
